Use CP 437 instead of 850.  From Wikipedia:
"Code page 850 is a code page that was used in western Europe, under systems such as DOS. It was also sometimes used on English DOS systems although 
CP437 was generally the default on those. It was largely replaced first with windows-1252 (often mislabeled as ISO-8859-1) and later by UCS-2 and 
finally UTF-16 (while the NT line was natively unicode from the start issues of development tool support and compatibility with windows 9x kept most 
applications on the 8 bit code pages). According to Microsoft, it is obsolete and unsupported."

In my own experience accessing a device shared via Samba from both Linux and Windows systems, I need to set the Linux system to cp437, not 850.  And 
using "obsolete and unsupported" codepages is not generally a good idea.

That being said, it might be better to find out why 1252 didn't work, since it should.  Then again, I believe I had tried 1252 before eventually 
getting to cp437 and had issues with it too.



git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdelibs@861244 283d02a7-25f6-0310-bc7c-ecb5cbfe19da