Revert Aaron's r664813 change (for kapplication.cpp only) to enable -style option on !X11 targets too.
Reviewed by Aaron.

Aaron's commit says
"only set the style when we're on x11; let Qt handle the style and palette selection on the other systems.
if people wish to change the style of Qt apps on those platforms, they can use qtconfig"

But from http://doc.trolltech.com/4.4/qtconfig.html:

"qtconfig is available on X11 platforms and should be installed alongside the other tools supplied with Qt."


I and possibly others need the quick --style option 
during the development for various style/appearance-related 
tests on the non-X11 targets.




git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdelibs@883540 283d02a7-25f6-0310-bc7c-ecb5cbfe19da