Code folding marker improvements: fix initial color, fix arrow luma, eliminate unneeded second color set.
This also removes the HSL blending code in favor of KCU::tint, which seems to work well. If not, we should add an HCY blend along the lines of ::tint (specifically, in how it handles the HC components for low-chroma colors, i.e. it should be ::tint but with linear Y response) to KCU.

TODO: make configurable!

BUG: 173695


git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdelibs@908568 283d02a7-25f6-0310-bc7c-ecb5cbfe19da