fixed: if KSelectAction gets disabled, its KComboBox isn't (workaround for QWidgetAction flaw)
KSelectAction has the problem that it does not disable its KComboBox, if is disabled itself.
It works if it is in MenuMode, because there KSelectAction creates a QToolButton and sets itself
as the default action. As a QToolButton reacts to changes of its default action it also syncs with
the enabled state. One would expect that the base class QWidgetAction does this itself also for all
other widgets, but obviously it doesn't (Qt 4.4.3). Reported to the Trolls.

Is there a smarter solution?
Okay to backport to 4.2?

CCMAIL:kde-core-devel@kde.org


git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdelibs@908643 283d02a7-25f6-0310-bc7c-ecb5cbfe19da