Merged revision 922955: Some fixes for our stylesheets collection to fix  http://www.webtoolkit.eu/wt:
1) Make sure that we create a stylesheet object even for empty <style> elements
2) Keep track of dirtiness of stylesheets list separately from that of the style selector,
and make sure to rebuild it if the collection is accessed. While I am at it, 
instead of having confusingly similarly named recalcStyleSelector and 
rebuildStyleSelector, have rebuildStyleSelector and rebuildStyleSheetList, 
the difference between which should be evident. Also make them private.
(Though updateStyleSelector is not renamed, so things aren't 100% clear, still)

BUG: 182230

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdelibs@922959 283d02a7-25f6-0310-bc7c-ecb5cbfe19da