Fix ruby syntax bugs
A single-quoted single-character string inside a shell command caused
the rest of the file to be incorrectly highlighted.

A general delimited input expression passed as the first argument to a
method without using parenthesis was incorrectly interpreted as modulo.

This bug was very common when using the puts method:
puts %(hello world)

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdelibs@935483 283d02a7-25f6-0310-bc7c-ecb5cbfe19da