Briefly: I added quite a few features to the right-mouse-button context menu.
* Added "Delete"
* Added "Move to Trash"
* Added a seperator somewhere in there...
* Added "Properties"

This is a move towards more similarity to the "Open File Dialog", it has some pretty useful features, so we should too.

So... yeah, you can do all of those groovy things via the context menu now.

I also renamed a bunch of double-plus stupidly named slots, consisting of things like _k_SlotUser1() .... who would have thought that meant "Create a New Folder" *shrug*. I mean, yeah, you have to tell KDialog to use the User1 button, but this is more internal than that, and it should make more sense to name it something un-stupid. Apparently this was previously used for KDialog compatibility, but no more shall it be.

...Think that's everything...

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdelibs@939110 283d02a7-25f6-0310-bc7c-ecb5cbfe19da