Qt < 4.5 had a bug in the layout of QAbstractScrollArea's scrollbars, that made impossible to properly compute the size of the viewport for styles using SH_ScrollView_FrameOnlyAroundContents (e.g. Oxygen).
This in turn caused all sorts of off-by-two artifacts.

We once had a symptomatic turnaround made with trial and errors, 
but it was wrong and ceased to work with recent Oxygen changes.
Here is the correct fix.

Please test with Qt < 4.4 on branch if you can. 

This is for backporting purposes only, and will be removed from trunk
in a moment.

BUG: 179359



git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdelibs@939487 283d02a7-25f6-0310-bc7c-ecb5cbfe19da