my previous attempt at merging continuations back into inlines in removeChild was way too shallow.
this ought to work for nested inlines too, so lets walk the chain
all the way back to the root anonymous block while merging.

this should cover most cases. To get a 100% correct behaviour, we will
have to get rid of continuations and use some sort of implicit inline
containers for hosting blocks in inlines.

CCBUG: 176681

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdelibs@942116 283d02a7-25f6-0310-bc7c-ecb5cbfe19da