CCBUG: 172306
With this patch, it's possible to cut ones and paste it multible times.
As the original files removed with the first paste, the clipboard is
updated with the new urls as a copy operation.

This patch doesn't change current cut-paste operation. So it has no
negative effect. Just exchange the cut-paste behavior as described in
the bug report.


git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdelibs@943265 283d02a7-25f6-0310-bc7c-ecb5cbfe19da