Made the "Add Entry..." action be displayed when clicked on items in the places view.
Before there would be a problem when you say... have limited space in the panel, but to add an entry, you would have to find that tiny 3 pixel-high space at the bottom, and click there, since you had to click on the panel itself, items would not give you access to this feature. 

I think this is bad for people with a.) disabilities b.) high-sensitive mouses or even c.) bad coordination ;-) 

So using RMB on places that are bookmarked shows this action there too; Logical, since it's near impossible to add entries when it's packed full.

I also added a separator between this entry (when clicked on items), and the "Hide" toggle action, since I see them as unrelated and, to me, makes sense to separate them.

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdelibs@945157 283d02a7-25f6-0310-bc7c-ecb5cbfe19da