replace the kdeglobals handling with something slightly less bizarre
KConfig("kdeglobals") would have essentially opened the file as a both
local and global file. this required lots of special casing and was
still broken.
so now just degrade it to a normal file with no globals. there is still
some special handling due to the ominous system.kdeglobals support, but
it is within reason.

CCMAIL: faure@kde.org


git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdelibs@946222 283d02a7-25f6-0310-bc7c-ecb5cbfe19da