Fixed regression introduced with SVN commit 948886: When the middle-mouse button has been clicked, 2 tabs get opened in Dolphin instead of one.
Connecting to the activated() signal should not be done in the context of KFilePlacesView. The behavior when the activated() signal is emitted depends on the used widget style, but the behavior of KFilePlacesView does not depend on the used style (-> no double click required to open a place, even if the widget style wants a double click behavior). 

CCMAIL: shaun.reich@kdemail.net

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdelibs@952197 283d02a7-25f6-0310-bc7c-ecb5cbfe19da