Really new password strength computation
Well, my last commit fixed the 12345 issue, but I was not
really satisfied. With this new algorithm you will really
have a hard time to satisfy 100% because it computes the
number of character category changes, and changes between
vowels and consonants are NOT counted, because this will
look like a word!



git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdelibs@952601 283d02a7-25f6-0310-bc7c-ecb5cbfe19da