-remove this file
This is no source incompatible change, because this file is 
a) not for building anything with cmake, but for setting up  Nightly builds (which nobody has done until now). 
b) It also was not in a working state at all, so removing it cannot break anyhting.

Alex

CCMAIL: kde-buildsystem@kde.org



git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdelibs@955472 283d02a7-25f6-0310-bc7c-ecb5cbfe19da