don't give focus to wrong view
don't use an anonymous single shot timer for managing the closing of the
command line editor. KateCmdLineEdit::hideLineEdit() will give focus to
it's m_view, which after changing views will be the wrong one.
instead, keep a qtimer pointer and connect the view's focusout signal to
the timer's stop slot.

BUG: 182398

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdelibs@961949 283d02a7-25f6-0310-bc7c-ecb5cbfe19da