Bugfix: do not load an empty/nonexistant config file for the spell checker in the Highlighter from KTextEdit
Example: my kde global locale is "en_US", by my global settings for spellchecking is "es"
I only have a spell checker for "es" (none for "en_US")

If we set the Spell Loader settings using an empty file, instead of using the global sonnetrc file
configurations (which properly setup "es" which is installed), it will use the KGlobal::locale lang, 
which is "en_US" (which is not installed neither), causing the Highlighter to not work at all.

Besides this, as the global static Loader object has now its settings modified, any other attempt
to launch the Spell Dialog will initiate the wrong SpellChecker and it won't work at all until we
restart the application or reset the KTextEdit spell config file.



git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdelibs@976805 283d02a7-25f6-0310-bc7c-ecb5cbfe19da