Reinitialize painter for ArrowDown pixmap
The interesting part about the backtrace (also in both duplicates)
is that the crash happens when rendering the "Down" arrow, although
the same code is used to render the first arrow. There might be a
problem in Qt with the state the painter is in, so reinitialize the
painter by correctly using begin() and end() calls.

CCBUG: 193183

@Praveen, as far as I know you were able to reproduce using the
xnest application. Could you test if this fixes the crash for you?



git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdelibs@977397 283d02a7-25f6-0310-bc7c-ecb5cbfe19da