-- When a KLineEdit is initialized with no text in the constructor, the KLineEditButton::animateVisible function is first called with false (animate to invisible), and the direction is set to Backward
When a text is set by code or by typing, animateVisible is called with true
and the button is properly shown.

However, when text is set on constructor of KLineEdit, the first animateVisible
call with false will not be done, but the one with true will do.
And it will return without showing the clear button because:
m_timeline->direction() == QTimeLine::Forward (as set by constructor)
--

Initialize KLineEditButton::m_timeLine as Backward will fix this issue, avoiding
the need of the animateVisible(false) call for this widget to work.

However, the button should have a better state of "I'm visible" to avoid
calling to "animate to invisible" at first time, when the button isn't even shown.

BUG: 193045



git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdelibs@982488 283d02a7-25f6-0310-bc7c-ecb5cbfe19da