Handle the cases where a directory (without filename) is passed to one of the KFileDialog::getSaveFile* functions.  Don't handle the startDir parameter to these (or any other of the static functions) in any way specially, just pass it straight in to the constructor.  This and KFileWidget will take care of parsing and checking the parameter.
Ensure that the KFile::LocalOnly option is set for the getSaveFile* functions.
Removed kWarning() for a non-local file in these - if anything is to be done
about non-local files then it should be in KFileWidget.

Also pass the parameter straight through for the KEncodingFileDialog static functions.

Reviewed at http://reviewboard.kde.org/r/803/

BUG:194900


git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdelibs@990505 283d02a7-25f6-0310-bc7c-ecb5cbfe19da