make mapping code mode agnostic.
move all mapping commands to kateviglobal and make them take an argument
describing which mode you want to set/get mappings for. this will make it
easier to add support for mapping in other modes later.

katevimodebase is now an abstract class, which it should have been from
the start.

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdelibs@994240 283d02a7-25f6-0310-bc7c-ecb5cbfe19da