Add the KReparentingProxyModel to kdelibs.
Only initial work has been done, but it's in kdelibs now so it can be used with
proxymodeltestapp and proxymodeltest. It is not built anywhere because it relies on
patches not merged into Qt.

This will probably be used for threading items in pim applications in the future,
using the mail threading agent to create item attributes to determine where they belong.

CCMAIL: kde-pim@kde.org

git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/KDE/kdelibs@1008438 283d02a7-25f6-0310-bc7c-ecb5cbfe19da